Update of match behavior when clicking on the icone
Users can send a new match proposal by clickning on the heart icon
Users can then unsend the demand by clicking again on the blue heart icon.

If another user accepts the demand, the icon becomes green and is no longer clickable.
If the user refuses the demand, the icon becomes red and is no longer clickable.

If two users send a match demand to each other, the demand is automatically accepted